Moves code out of lambda file. Shares responses with other lambdas.
