+++ api/lib/term-synonyms.test.ts
@@ -1,0 +1,3 @@
+describe('term-synonyms', () => {
+  test.todo('getSynonyms');
+});
